Renamed test directories
