+++ code_tests/integration_tests/test_adjacent_news_api.py
