+++ code_tests/integration_tests/test_benchmarker.py
@@ -5,7 +5,7 @@ from unittest.mock import Mock
 
 import pytest
 
-from code_tests.unit_tests.test_forecasting.forecasting_test_manager import (
+from code_tests.unit_tests.forecasting_test_manager import (
     ForecastingTestManager,
     MockBot,
 )
