+++ code_tests/integration_tests/test_forecast_bots_live.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock
 import pytest
 import typeguard
 
-from code_tests.unit_tests.test_forecasting.forecasting_test_manager import (
+from code_tests.unit_tests.forecasting_test_manager import (
     ForecastingTestManager,
 )
 from forecasting_tools.ai_models.general_llm import GeneralLlm
