+++ code_tests/integration_tests/test_forecast_database_manager.py
