+++ code_tests/integration_tests/test_smart_searcher.py
