+++ code_tests/integration_tests/test_structure_output.py
