+++ code_tests/unit_tests/test_auto_optimizers/test_question_plus_research.py
@@ -1,6 +1,6 @@
 import pytest
 
-from code_tests.unit_tests.test_forecasting.forecasting_test_manager import (
+from code_tests.unit_tests.forecasting_test_manager import (
     ForecastingTestManager,
 )
 from forecasting_tools.auto_optimizers.question_plus_research import (
